% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_index_calculation.R
\name{ai_index_calculation}
\alias{ai_index_calculation}
\title{Calculates the aridity index}
\usage{
ai_index_calculation(annual_img_ppt, annual_img_etp)
}
\arguments{
\item{annual_img_ppt}{rast image with the annual sum of the precipitation}

\item{annual_img_etp}{rast image of annual sum of the potential evapotranspiration}
}
\value{
Returns a SpatRaster with the calculation.
}
\description{
The aridity index (AI) is a numerical indicator used to quantify the degree of dryness of a climate at a given location. It is calculated as the ratio of annual precipitation to potential evapotranspiration.
}
\examples{
\dontrun{

 ai_index_img <- ai_index_calculation(annual_img_ppt, annual_img_etp);
}
}
