% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_ppt_etp.R
\name{etp_calibration}
\alias{etp_calibration}
\title{Calibration of minimum air temperature from TerraClimate}
\usage{
etp_calibration(slope, intercept, etp_stack)
}
\arguments{
\item{slope}{the slope of the linear regression (numeric).}

\item{intercept}{the intercept of the linear regression (numeric).}

\item{etp_stack}{stack of etp.}
}
\value{
Returns a Rasterstack of etp calibrated.
}
\description{
This fuction will calibrate the reference evapotranspiration (etp) from TerraClimate dataset based in the relationship with observed weather stations data.
}
\examples{
\dontrun{
etp_cal<- etp_calibration(slope = 0.930073,
                           intercept = 22.399986, etp_stack = etp);
}
}
