% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_of_annual_sum_of_variables.R
\name{mean_of_annual_sum_of_variables}
\alias{mean_of_annual_sum_of_variables}
\title{Calculates the mean of the annual sum of the variables values}
\usage{
mean_of_annual_sum_of_variables(list_dir_of_imgs)
}
\arguments{
\item{list_dir_of_imgs}{directory containing the images downloaded in download_terraclimate}
}
\value{
Returns a SpatRaster with the calculation.
}
\description{
The function calculates the mean of the annual sum of the variables of interest.
}
\examples{
\dontrun{

annual_mean_ppt<- mean_of_annual_sum_of_variables(list_dir_of_imgs = C:/ppt_dir);
}
}
