% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AWC.R
\name{plot_AWC}
\alias{plot_AWC}
\title{Function to plot the percentage of Available Water Capacity (AWC)
\if{html}{\figure{logo_cropDemand.png}{options: height= 300 width=auto style = float:right alt= Our logo}}}
\usage{
plot_AWC(AWC_stack)
}
\arguments{
\item{AWC_stack}{A stack generated in WaterDemand function}
}
\value{
Returns a  plot (gg file) of monthly percentage of AWC
}
\description{
This function will plot the monthly AWC
}
\examples{
\dontrun{
plot_AWC(AWC_stack)
}
}
