% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_brazil_biomes.R
\name{see_brazil_biomes}
\alias{see_brazil_biomes}
\title{Function to see the Brazilian biomes available tp use in download.terraclimate and loadROI function}
\usage{
see_brazil_biomes()
}
\value{
The biomes information available to run the cropDemand package.
}
\description{
This function will show the biomes available in the package and how we can call each biome polygon.
}
\examples{
\dontrun{
see_brazil_biomes()
}
}
