% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_totram.R
\name{w_totram}
\alias{w_totram}
\title{Soil moisture estimation based in TOTRAM's models}
\usage{
w_totram(img_ndvi, img_lst)
}
\arguments{
\item{img_ndvi}{rasterstack of NDVI (Normalized Difference Vegetation Index)}

\item{img_lst}{rasterstack of LST (Land Surface Temperature)}
}
\value{
Returns a Rasterstack of soil moisture (W).
}
\description{
This fuction will calculate the soil moisture based in TOTRAM's models, which uses NDVI and LST as input (Rasterstack).
}
